+++ ep-sdk/src/tasks/EpSdkApplicationVersionTask.ts
@@ -15,10 +15,10 @@ import {
   IEpSdkTask_Keys, 
   IEpSdkTask_UpdateFuncReturn 
 } from './EpSdkTask';
-import { EpSdkVersionTask, IEpSdkVersionTask_Config, IEpSdkVersionTask_EpObjectKeys } from './EpSdkVersionTask';
+import { EEpSdk_VersionTaskStrategy, EpSdkVersionTask, IEpSdkVersionTask_Config, IEpSdkVersionTask_EpObjectKeys } from './EpSdkVersionTask';
 
-type TEpSdkApplicationVersionTask_Settings = Required<Pick<ApplicationVersion, "description" | "displayName" | "stateId" >>;
-type TEpSdkApplicationVersionTask_CompareObject = Partial<TEpSdkApplicationVersionTask_Settings>;
+export type TEpSdkApplicationVersionTask_Settings = Required<Pick<ApplicationVersion, "description" | "displayName" | "stateId" >>;
+type TEpSdkApplicationVersionTask_CompareObject = Partial<TEpSdkApplicationVersionTask_Settings> & Pick<ApplicationVersion, "version">;
 
 export interface IEpSdkApplicationVersionTask_Config extends IEpSdkVersionTask_Config {
   applicationDomainId: string;
@@ -139,14 +139,20 @@ export class EpSdkApplicationVersionTask extends EpSdkVersionTask {
     }}));
 
     if(epSdkApplicationVersionTask_GetFuncReturn.epObject === undefined) throw new EpSdkInternalTaskError(logName, this.constructor.name, 'epSdkApplicationVersionTask_GetFuncReturn.epObject === undefined');
-
+    if(epSdkApplicationVersionTask_GetFuncReturn.epObject.version === undefined) throw new EpSdkApiContentError(logName, this.constructor.name, 'epSdkApplicationVersionTask_GetFuncReturn.epObject.version === undefined', {
+      epObject: epSdkApplicationVersionTask_GetFuncReturn.epObject
+    });
     const existingObject: ApplicationVersion = epSdkApplicationVersionTask_GetFuncReturn.epObject;
     const existingCompareObject: TEpSdkApplicationVersionTask_CompareObject = {
       description: existingObject.description,
       displayName: existingObject.displayName,
-      stateId: existingObject.stateId
+      stateId: existingObject.stateId,
     };
     const requestedCompareObject: TEpSdkApplicationVersionTask_CompareObject = this.createObjectSettings();
+    if(this.versionStrategy === EEpSdk_VersionTaskStrategy.EXACT_VERSION) {
+      existingCompareObject.version = epSdkApplicationVersionTask_GetFuncReturn.epObject.version;
+      requestedCompareObject.version = this.versionString;
+    }
 
     const epSdkTask_IsUpdateRequiredFuncReturn: IEpSdkTask_IsUpdateRequiredFuncReturn = this.create_IEpSdkTask_IsUpdateRequiredFuncReturn({ 
       existingObject: existingCompareObject,