+++ ep-sdk/test/specs/tasks/applicationVersionTask.spec.ts
@@ -20,11 +20,13 @@ import {
   TEpSdkVersionTaskStrategyValidationError_Details
 } from '../../../src/utils/EpSdkErrors';
 import { EEpSdkTask_Action, EEpSdkTask_TargetState } from '../../../src/tasks/EpSdkTask';
-import { EpSdkApplicationVersionTask, IEpSdkApplicationVersionTask_ExecuteReturn } from '../../../src/tasks/EpSdkApplicationVersionTask';
+import { EpSdkApplicationVersionTask, IEpSdkApplicationVersionTask_ExecuteReturn, TEpSdkApplicationVersionTask_Settings } from '../../../src/tasks/EpSdkApplicationVersionTask';
 import EpSdkStatesService from '../../../src/services/EpSdkStatesService';
 import EpSdkApplicationDomainsService from '../../../src/services/EpSdkApplicationDomainsService';
 import { EEpSdk_VersionTaskStrategy } from '../../../src/tasks/EpSdkVersionTask';
 import { IEpSdkTask_TransactionLogData } from '../../../src/tasks/EpSdkTask_TransactionLog';
+import EpSdkSemVerUtils, { EEpSdk_VersionStrategy } from '../../../src/utils/EpSdkSemVerUtils';
+import EpSdkApplicationVersionsService from '../../../src/services/EpSdkApplicationVersionsService';
 
 
 const scriptName: string = path.basename(__filename);
@@ -432,6 +434,126 @@ describe(`${scriptName}`, () => {
     }
   });
 
+  it(`${scriptName}: application version present: create version without updates to settings`, async () => {
+    const settings: TEpSdkApplicationVersionTask_Settings = {
+      stateId: EpSdkStatesService.releasedId,
+      description: 'description',
+      displayName: 'displayName',
+    }
+    try {
+      // create a reference version
+      const epSdkApplicationVersionTask_Ref = new EpSdkApplicationVersionTask({
+        epSdkTask_TargetState: EEpSdkTask_TargetState.PRESENT,
+        applicationDomainId: ApplicationDomainId,
+        applicationId: ApplicationId,
+        // versionString: newVersion,
+        // versionStrategy: EEpSdk_VersionTaskStrategy.EXACT_VERSION,
+        applicationVersionSettings: settings,
+        epSdkTask_TransactionConfig: {
+          parentTransactionId: 'parentTransactionId',
+          groupTransactionId: 'groupTransactionId'
+        },
+        checkmode: false,
+      });
+      const epSdkApplicationVersionTask_ExecuteReturn_Ref: IEpSdkApplicationVersionTask_ExecuteReturn = await epSdkApplicationVersionTask_Ref.execute();
+      const referenceVersionString: string = epSdkApplicationVersionTask_ExecuteReturn_Ref.epObject.version ? epSdkApplicationVersionTask_ExecuteReturn_Ref.epObject.version : 'not-a-version';
+      // bump the version
+      const newVersion = EpSdkSemVerUtils.createNextVersionByStrategy({
+        fromVersionString: referenceVersionString,
+        strategy: EEpSdk_VersionStrategy.BUMP_MINOR,
+      });
+      // create new version even no other updates to settings
+      const epSdkApplicationVersionTask_New = new EpSdkApplicationVersionTask({
+        epSdkTask_TargetState: EEpSdkTask_TargetState.PRESENT,
+        applicationDomainId: ApplicationDomainId,
+        applicationId: ApplicationId,
+        versionString: newVersion,
+        // versionStrategy: EEpSdk_VersionTaskStrategy.EXACT_VERSION,
+        applicationVersionSettings: settings,
+        epSdkTask_TransactionConfig: {
+          parentTransactionId: 'parentTransactionId',
+          groupTransactionId: 'groupTransactionId'
+        },
+        checkmode: false,
+      });
+      let epSdkApplicationVersionTask_ExecuteReturn: IEpSdkApplicationVersionTask_ExecuteReturn = await epSdkApplicationVersionTask_New.execute();
+      let message = TestLogger.createLogMessage('epSdkApplicationVersionTask_ExecuteReturn', epSdkApplicationVersionTask_ExecuteReturn);
+      // get the latest version to check
+      let latestVersionString = await EpSdkApplicationVersionsService.getLatestVersionString({ applicationId: ApplicationId });
+      // expect no change in version
+      expect(latestVersionString, message).to.eq(referenceVersionString);
+      expect(epSdkApplicationVersionTask_ExecuteReturn.epSdkTask_TransactionLogData.epSdkTask_Action, message).to.eq(EEpSdkTask_Action.NO_ACTION);
+
+      // now test exact match, checkmode=true for newVersion: should return Would create new version
+      const epSdkApplicationVersionTask_NewCreatedCheck = new EpSdkApplicationVersionTask({
+        epSdkTask_TargetState: EEpSdkTask_TargetState.PRESENT,
+        applicationDomainId: ApplicationDomainId,
+        applicationId: ApplicationId,
+        versionString: newVersion,
+        versionStrategy: EEpSdk_VersionTaskStrategy.EXACT_VERSION,
+        applicationVersionSettings: settings,
+        epSdkTask_TransactionConfig: {
+          parentTransactionId: 'parentTransactionId',
+          groupTransactionId: 'groupTransactionId'
+        },
+        checkmode: true,
+      });
+      epSdkApplicationVersionTask_ExecuteReturn = await epSdkApplicationVersionTask_NewCreatedCheck.execute();
+      message = TestLogger.createLogMessage('epSdkApplicationVersionTask_ExecuteReturn', epSdkApplicationVersionTask_ExecuteReturn);
+      expect(epSdkApplicationVersionTask_ExecuteReturn.epSdkTask_TransactionLogData.epSdkTask_Action, message).to.eq(EEpSdkTask_Action.WOULD_CREATE_NEW_VERSION);
+      expect(epSdkApplicationVersionTask_ExecuteReturn.epObject.version, message).to.eq(newVersion);
+
+      // now test exact match, checkmode=false for newVersion: should create it
+      const epSdkApplicationVersionTask_NewCreated = new EpSdkApplicationVersionTask({
+        epSdkTask_TargetState: EEpSdkTask_TargetState.PRESENT,
+        applicationDomainId: ApplicationDomainId,
+        applicationId: ApplicationId,
+        versionString: newVersion,
+        versionStrategy: EEpSdk_VersionTaskStrategy.EXACT_VERSION,
+        applicationVersionSettings: settings,
+        epSdkTask_TransactionConfig: {
+          parentTransactionId: 'parentTransactionId',
+          groupTransactionId: 'groupTransactionId'
+        },
+        checkmode: false,
+      });
+      epSdkApplicationVersionTask_ExecuteReturn = await epSdkApplicationVersionTask_NewCreated.execute();
+      message = TestLogger.createLogMessage('epSdkApplicationVersionTask_ExecuteReturn', epSdkApplicationVersionTask_ExecuteReturn);
+      expect(epSdkApplicationVersionTask_ExecuteReturn.epSdkTask_TransactionLogData.epSdkTask_Action, message).to.eq(EEpSdkTask_Action.CREATE_NEW_VERSION);
+      expect(epSdkApplicationVersionTask_ExecuteReturn.epObject.version, message).to.eq(newVersion);
+      latestVersionString = await EpSdkApplicationVersionsService.getLatestVersionString({ applicationId: ApplicationId });
+      expect(latestVersionString, message).to.eq(newVersion);
+
+      // now test exact match, checkmode=true going back to reference version: should return Would fail
+      const epSdkApplicationVersionTask_RefCheck = new EpSdkApplicationVersionTask({
+        epSdkTask_TargetState: EEpSdkTask_TargetState.PRESENT,
+        applicationDomainId: ApplicationDomainId,
+        applicationId: ApplicationId,
+        versionString: referenceVersionString,
+        versionStrategy: EEpSdk_VersionTaskStrategy.EXACT_VERSION,
+        applicationVersionSettings: settings,
+        epSdkTask_TransactionConfig: {
+          parentTransactionId: 'parentTransactionId',
+          groupTransactionId: 'groupTransactionId'
+        },
+        checkmode: true,
+      });
+      epSdkApplicationVersionTask_ExecuteReturn = await epSdkApplicationVersionTask_RefCheck.execute();
+      message = TestLogger.createLogMessage('epSdkApplicationVersionTask_ExecuteReturn', epSdkApplicationVersionTask_ExecuteReturn);
+      expect(epSdkApplicationVersionTask_ExecuteReturn.epSdkTask_TransactionLogData.epSdkTask_Action, message).to.eq(EEpSdkTask_Action.WOULD_FAIL_CREATE_NEW_VERSION_ON_EXACT_VERSION_REQUIREMENT);
+      expect(epSdkApplicationVersionTask_ExecuteReturn.epObject.version, message).to.eq(referenceVersionString);
+      latestVersionString = await EpSdkApplicationVersionsService.getLatestVersionString({ applicationId: ApplicationId });
+      expect(latestVersionString, message).to.eq(newVersion);
+
+      // could also check the error for checkmode = false
+      // DEBUG
+      // expect(false, message).to.be.true;
+    } catch(e) {
+      if(e instanceof ApiError) expect(false, TestLogger.createApiTestFailMessage('failed')).to.be.true;
+      expect(e instanceof EpSdkError, TestLogger.createNotEpSdkErrorMessage(e)).to.be.true;
+      expect(false, TestLogger.createEpSdkTestFailMessage('failed', e)).to.be.true;
+    }
+  });
 
 });